app/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coverage==7.8.0
cryptography==44.0.2
defusedxml==0.7.1
Django==4.2.20
django-allauth==0.58.2
django-cors-headers==4.7.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
idna==3.10
oauthlib==3.2.2
psycopg==3.1.20
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
sqlparse==0.5.3
typing_extensions==4.13.0
urllib3==2.3.0
